Create labelImg_pascalVOC_to_darkflow.py
This script  updates annotation xml files that were created with labelImg to be compatible with this fork of the darkflow implementation.
The labelImg VOC/Pascal format save the filename extension as xml. Darkflow use this 
field as the image file name when it parse and batch the data for training 
(see parse and _batch function in darkflow/darkflow/yolo/data.py and 
darkflow/darkflow/utils/pascal_voc_clean_xml.py).